Fix test fail and other uninitialized cfloat errors and warnings
 * Fix cfloat_special_cases iszero() test that FAILed in regression
 * Fix build warnings (with gcc12) of cfloat 'may be used uninitialized'
 * Fix an unrelated use of an uninitialized variable in rpoly_ak1.cpp
 * Add initializations for many uninitialized cfloats
 * Add 'uninitialized' comments on other uninitialized cfloats

The initial PR commit makes cfloat a trivial type.
This commit is a reminder that this *** IS A BREAKING CHANGE ***.

The regression test failure was a lucky break - now a flaky test it can
PASS or FAIL depending on whether memory is zero-initialized or not.
The test case now performs placement-new to a garbage filled buffer,
both expected-fail for uninitialized and expected-pass when initialized.

A consequence of being a trivial type is that uninitialized declarations
are truly uninitialized so can have garbage value, depending on whatever
their storage held previously.

There are many such uninitialized cfloats throughout the universal repo.
In the implementation this can be a performance feature, where necessary.
However, it is rarely necessary to declare a variable uninitialized and
later initialize it. Guideline: "Don’t use two-phase initialization".

This PR adds initializations for many previously unitialized cfloats, or
it adds a comment to remind that they are declared as uninitialized:

  Cfloat a; // uninitialized
  Cfloat a{};

In many cases typedef aliases are used, so it is likely that there are
many more cases that were missed (but probably safe if gcc doesn't warn).

In fixing these, I found another warning on variable 'omp' being used
before being initialized, in rpoly_ak1.cpp - I fix it here
(not worth another PR, adding the init is a safe change).

Material for Release Notes

The next release of universal should flag up this change to trivial types
as a potential breaking change. Here are some useful resources.

Make sure to enable all compiler warnings for uninitialized variables,
and be sure to look at and heed the warnings.

https://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines#Rnr-two-phase-init

https://clang.llvm.org/docs/AttributeReference.html#uninitialized

A last resort is to use a flag to initialize auto vars -ftrivial-auto-var-init
https://reviews.llvm.org/D54604?id=174471

https://devblogs.microsoft.com/cppblog/new-code-analysis-quick-fixes-for-
uninitialized-memory-c6001-and-use-before-init-c26494-warnings/

Remember; the importance of being trivial is improved performance in APIs.
Universal types like cfloat should be trivial, so users must navigate the
perils of triviality to enjoy the benefits.