Apply noexcept consistently
Universal has a lot of noexcept spec functions, and that seems the right
decision for a provider of low level types. The noexcept usage in this
blocksignificant file was inconsistent, so I made it more consistent by
adding noexcept in most places, apart from on the division functions in
case divide by zero is defined as throwing in future.